Clean up web server to put more logic into middlewares.
It's no longer a single giant behaviour soup main file.
