+++ web/server.js
@@ -1,17 +1,11 @@
 'use strict'
 
-var fs = require('fs')
 var path = require('path')
 var http = require('http')
 var express = require('express')
 var compression = require('compression')
 var morgan = require('morgan')
-var handlebars = require('handlebars')
-
-var clientConfig = require('./clientConfig')
-var indexContext = {
-  clientConfig: JSON.stringify(clientConfig),
-}
+var clientConfig = require('./client-config')
 
 var app = express()
 
@@ -24,20 +18,11 @@ if (!process.env.PORT) {
 }
 
 // Serve assets right out of the dist tree.
-app.use('/bundle', express.static('dist/bundle', { maxAge: '1 year' }))
-app.use('/assets', express.static('dist/assets'))
+app.use('/bundle', express.static(path.resolve('dist/bundle'), { maxAge: '1 year' }))
 
 // HTML5 pushstate routing support. By the time we get here, we definitely did
 // not match an asset. Serve the index.html document, instead.
-app.get(/.*/, function(req, res) {
-  // We could precompile this, but live reload is nice in development.
-  fs.readFile('dist/index.html', 'UTF-8', function(err, template) {
-    if (err) throw err
-    var renderer = handlebars.compile(template)
-    var body = renderer(indexContext)
-    res.send(body)
-  })
-})
+app.get(/.*/, clientConfig(path.resolve('dist/index.html')))
 
 var port = process.env.PORT || 4000
 app.listen(port, function() {
